修改字段转换

+++ TableToStruct.go
@@ -45,7 +45,7 @@ var typeForMysqlToGo = map[string]string{
 	"datetime":           "time.Time", // time.Time or string
 	"timestamp":          "time.Time", // time.Time or string
 	"time":               "time.Time", // time.Time or string
-	"float":              "float64",
+	"float":              "float32",
 	"double":             "float64",
 	"decimal":            "float64",
 	"binary":             "string",
